requirements.txt
streamlit
pandas
numpy
matplotlib
snowflake-connector-python
snowflake-snowpark-python

google-generative-ai
json
os
time

dotenv

typing
re

scikit-learn
math
fuzzywuzzy
plotly

uuid
